Properly broke up schemas
